OpenVoxel: Work on proper implementation of a world cull for rendering

+++ OpenVoxel/src/main/java/net/openvoxel/world/client/ClientChunk.java
@@ -18,8 +18,8 @@ public class ClientChunk extends Chunk {
 		}
 	}
 
-	public ClientChunkSection getSectionAt(int y) {
-		return (ClientChunkSection)chunkSections[y];
+	public ClientChunkSection getSectionAt(long y) {
+		return (ClientChunkSection)chunkSections[(int)y];
 	}
 
 }
